requirements.txt
joblib==1.4.2
boto3==1.35.0
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.6.1 